Refactored request validator
