+++ Service/src/request_validator/schemas/charts_schema.py
@@ -1,12 +1,5 @@
 from ..field import Field
-
-def validateLength(lengthStr, field, errors):
-   if type(lengthStr) != str or lengthStr not in ['day', 'week', 'month', 'year', 'recent']:
-      errors.append({
-         'InvalidField': field.name + " is not one of 'day', 'week', 'month', 'year', or 'recent'"
-      })
-   
-   return errors
+from .util_schema import validateLength
 
 fields = [
    Field('ticker', str, True),
