+++ Service/src/request_validator/schemas/portfolio_schema.py
@@ -1,0 +1,9 @@
+from ..field import Field
+from .util_schema import validateName
+
+fields = [
+   Field('name', str, True, validateName),
+   Field('buyPower', int, False),
+   Field('inviteCode', str, False)
+]
+
